+++ cms-20240330/alibabacloud_cms20240330/models/_get_memory_response_body.py
@@ -7,6 +7,8 @@ from darabonba.model import DaraModel
 class GetMemoryResponseBody(DaraModel):
     def __init__(
         self,
+        agent_id: str = None,
+        app_id: str = None,
         created_at: str = None,
         expiration_date: str = None,
         id: str = None,
@@ -16,8 +18,12 @@ class GetMemoryResponseBody(DaraModel):
         organization: str = None,
         owner: str = None,
         request_id: str = None,
+        run_id: str = None,
         updated_at: str = None,
+        user_id: str = None,
     ):
+        self.agent_id = agent_id
+        self.app_id = app_id
         self.created_at = created_at
         self.expiration_date = expiration_date
         self.id = id
@@ -27,7 +33,9 @@ class GetMemoryResponseBody(DaraModel):
         self.organization = organization
         self.owner = owner
         self.request_id = request_id
+        self.run_id = run_id
         self.updated_at = updated_at
+        self.user_id = user_id
 
     def validate(self):
         pass
@@ -37,6 +45,12 @@ class GetMemoryResponseBody(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.agent_id is not None:
+            result['agentId'] = self.agent_id
+
+        if self.app_id is not None:
+            result['appId'] = self.app_id
+
         if self.created_at is not None:
             result['createdAt'] = self.created_at
 
@@ -64,13 +78,25 @@ class GetMemoryResponseBody(DaraModel):
         if self.request_id is not None:
             result['requestId'] = self.request_id
 
+        if self.run_id is not None:
+            result['runId'] = self.run_id
+
         if self.updated_at is not None:
             result['updatedAt'] = self.updated_at
+
+        if self.user_id is not None:
+            result['userId'] = self.user_id
 
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('agentId') is not None:
+            self.agent_id = m.get('agentId')
+
+        if m.get('appId') is not None:
+            self.app_id = m.get('appId')
+
         if m.get('createdAt') is not None:
             self.created_at = m.get('createdAt')
 
@@ -98,8 +124,14 @@ class GetMemoryResponseBody(DaraModel):
         if m.get('requestId') is not None:
             self.request_id = m.get('requestId')
 
+        if m.get('runId') is not None:
+            self.run_id = m.get('runId')
+
         if m.get('updatedAt') is not None:
             self.updated_at = m.get('updatedAt')
+
+        if m.get('userId') is not None:
+            self.user_id = m.get('userId')
 
         return self